http2-app-flood/shared/requirements.txt
# HTTP/2 Flood Lab - All Requirements

# Core web server and attack components
quart==0.19.4
hypercorn==0.14.4
httpx[http2]==0.25.2

# All scenario dependencies
Pillow==10.1.0
cryptography==41.0.7